Changed check auth code, login and home page protected routes.
I created two components, NeedGuestPage and NeedLoginPage
NeedGuestPage is for pages where it is necessary the user
is not logged in (for example login page)

NeedLoginPage is for pages that require Login.

I moved the code inside onAuthStateChanged to a file
setUserReduxStore.ts to set all the data to the redux-toolkit
state.